4.22

+++ part4/bloglist/controllers/blogs.js
@@ -1,7 +1,6 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware');
 
 blogRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({})
@@ -9,14 +8,10 @@ blogRouter.get('/', async (request, response) => {
       
   })
   
-blogRouter.post('/', async (request, response) => {
+blogRouter.post('/', middleware.userExtractor, async (request, response) => {
     const body = request.body
 
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!decodedToken.id) {
-      return response.status(401).json({ error: 'token invalid' })
-    }
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
 
     const blog = new Blog({
         title: body.title,
@@ -40,9 +35,17 @@ blogRouter.get('/:id', async (request, response) => {
 })
 
 //4.13
-blogRouter.delete('/:id', async (request, response) => {
+blogRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === user.id){
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
+    } else {
+        // handle the unauthorized case
+        response.status(403).json({ error: 'you are not authorized to delete this blog' })
+    }
 })
 
 //4.14
